chore(http): - examples are splitted for better intro to modkit-auth - Cargo.toml is cleaned from redundant sections - Unknown HTTP error is returned now
